# storage/api-docs/api-docs-admin.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Admin application",
        "description": "API endpoints for the admin application.",
        "contact": {
            "name": "viracoding",
            "email": "viracodingGplus@gmail.com"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/admin",
            "description": "development Server"
        }
    ],
    "paths": {
...
        "/courses": {
            "get": {
                "tags": [
                    "Course"
                ],
                "summary": "Get courses with pagination",
                "operationId": "e50bc90e7d3037855a77f65b1030b366",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "category_id query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Course"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Course"
                ],
                "summary": "Store a new Course",
                "operationId": "03145d7c7525d28b7924670b86f32223",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "category_id",
                                    "author",
                                    "description",
                                    "price",
                                    "image"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Course Title"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "format": "int64",
                                        "example": 1
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Course description"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50.99
                                    },
                                    "discount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10
                                    },
                                    "session": {
                                        "type": "integer",
                                        "format": "int32",
                                        "example": 20
                                    },
                                    "summary": {
                                        "type": "string",
                                        "example": "Course summary"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Course created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Course"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "با موفقیت ایجاد شد ."
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "$ref": "#/components/schemas/ValidationError"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/courses/{id}/image": {
            "post": {
                "tags": [
                    "Course"
                ],
                "summary": "Update an existing course image",
                "operationId": "69e19b3416887de68cc25541fc436d1a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the course",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "با موفقیت بروزرسانی شد."
                                        },
                                        "avatar": {
                                            "type": "string",
                                            "example": "http://example.com/storage/course/course_1.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or image not provided",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "تصویری ارسال نشده است"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "description": "URL to the first page",
                        "type": "string"
                    },
                    "last": {
                        "description": "URL to the last page",
                        "type": "string"
                    },
                    "prev": {
                        "description": "URL to the previous page",
                        "type": "string"
                    },
                    "next": {
                        "description": "URL to the next page",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "title": "ValidationError",
                "description": "خطای اعتبارسنجی",
                "properties": {
                    "data": {
                        "type": "object"
                    },
                    "status": {
                        "type": "integer",
                        "example": 404
                    },
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "دسته والد وجود ندارد!"
                    },
                    "errors": {
                        "type": "array",
                        "items": {}
                    }
                },
                "type": "object"
            },
            "Course": {
                "title": "Course",
                "description": "Represents a course.",
                "required": [
                    "title",
                    "hours",
                    "category_id",
                    "author",
                    "description",
                    "price",
                    "session"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier for the course.",
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "description": "The title of the course.",
                        "type": "string"
                    },
                    "hours": {
                        "description": "The number of hours for the course.",
                        "type": "number"
                    },
                    "category_id": {
                        "description": "The ID of the category to which the course belongs.",
                        "type": "integer",
                        "format": "int64"
                    },
                    "author": {
                        "description": "The author of the course.",
                        "type": "string"
                    },
                    "description": {
                        "description": "A brief description of the course.",
                        "type": "string"
                    },
                    "price": {
                        "description": "The price of the course.",
                        "type": "number"
                    },
                    "discount": {
                        "description": "The discount applied to the course.",
                        "type": "number"
                    },
                    "session": {
                        "description": "The number of sessions for the course.",
                        "type": "integer",
                        "format": "int64"
                    },
                    "summary": {
                        "description": "A summary of the course content.",
                        "type": "string"
                    },
                    "image": {
                        "description": "The image URL for the course.",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}